requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
paho-mqtt==1.3.1
pkg-resources==0.0.0
requests==2.18.4
urllib3==1.22
